suzerain: #1273 move store(state) to common
git-svn-id: https://svn.ices.utexas.edu/repos/pecos/suzerain/trunk@19024 0bd85f0b-c151-43b9-a1e9-27231104551f

+++ apps/channel_mean.cpp
@@ -93,7 +93,7 @@ static bool process(const char * filename)
     load(h.get(), scenario);
     load(h.get(), grid);
     load(h.get(), bspw);
-    assert(static_cast<unsigned>(bspw->ndof()) == grid.N.y());
+    assert(bspw->ndof() == grid.N.y());
 
     // Load zero-zero mode coefficients for all state variables
     Eigen::ArrayXr s_coeffs(grid.N.y(), field_names.size());
